refactor: make rockcraft-pack follow new call conventions
