+++ rockcraft/pack/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@craft-actions/rockcraft-pack",
+  "version": "0.0.1",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "build": "pnpm node ../../build.js src/index.ts dist/index.js",
+    "test": "vitest run"
+  },
+  "dependencies": {
+    "@actions/core": "^1.10.1",
+    "@craft-actions/common": "workspace:*"
+  }
+}
