+++ tsconfig.json
@@ -10,5 +10,6 @@
     "skipLibCheck": true,
     "types": ["node"]
   },
-  "include": ["src/**/*.ts", "tests/**/*.ts"]
+  "include": ["common/**/*.ts", "rockcraft/pack/**/*.ts"],
+  "exclude": ["**/dist/**"]
 }
